cast to an int

+++ scripts/learnphp.php
@@ -32,7 +32,7 @@ echo "The value of pi is " . PI;
 echo "</br>5 + 2 = " . (5+2);
 echo "</br>5 - 2 = " . (5-2);
 echo "</br>5 * 2 = " . (5*2);
-echo "</br>5 / 2 = " . (5/2);
+echo "</br>5 / 2 = " . (integer)(5/2);
 echo "</br>5 % 2 = " . (5%2);
 ?>
 </body>
